move a few string related functions to a dedicated file
