+++ src/str.c
@@ -1,0 +1,99 @@
+#include <string.h>
+
+#include "str.h"
+
+#define MAX_DIGITS_IN_UINT64 27
+
+bool adjustDecimals(char *src, uint32_t srcLength, char *target,
+                    uint32_t targetLength, uint8_t decimals) {
+    uint32_t startOffset;
+    uint32_t lastZeroOffset = 0;
+    uint32_t offset = 0;
+    if ((srcLength == 1) && (*src == '0')) {
+        if (targetLength < 2) {
+                return false;
+        }
+        target[0] = '0';
+        target[1] = '\0';
+        return true;
+    }
+    if (srcLength <= decimals) {
+        uint32_t delta = decimals - srcLength;
+        if (targetLength < srcLength + 1 + 2 + delta) {
+            return false;
+        }
+        target[offset++] = '0';
+        target[offset++] = '.';
+        for (uint32_t i = 0; i < delta; i++) {
+            target[offset++] = '0';
+        }
+        startOffset = offset;
+        for (uint32_t i = 0; i < srcLength; i++) {
+            target[offset++] = src[i];
+        }
+        target[offset] = '\0';
+    } else {
+        uint32_t sourceOffset = 0;
+        uint32_t delta = srcLength - decimals;
+        if (targetLength < srcLength + 1 + 1) {
+            return false;
+        }
+        while (offset < delta) {
+            target[offset++] = src[sourceOffset++];
+        }
+        if (decimals != 0) {
+            target[offset++] = '.';
+        }
+        startOffset = offset;
+        while (sourceOffset < srcLength) {
+            target[offset++] = src[sourceOffset++];
+        }
+  target[offset] = '\0';
+    }
+    for (uint32_t i = startOffset; i < offset; i++) {
+        if (target[i] == '0') {
+            if (lastZeroOffset == 0) {
+                lastZeroOffset = i;
+            }
+        } else {
+            lastZeroOffset = 0;
+        }
+    }
+    if (lastZeroOffset != 0) {
+        target[lastZeroOffset] = '\0';
+        if (target[lastZeroOffset - 1] == '.') {
+                target[lastZeroOffset - 1] = '\0';
+        }
+    }
+    return true;
+}
+
+char *u64str(uint64_t v)
+{
+  static char buf[27];
+
+  char *p = &buf[sizeof(buf)];
+  *(--p) = '\0';
+
+  if (v == 0) {
+    *(--p) = '0';
+    return p;
+  }
+
+  while (v > 0) {
+    *(--p) = '0' + (v % 10);
+    v = v/10;
+  }
+
+  return p;
+}
+
+char* amount_to_str(uint64_t amount, uint8_t decimals){
+  char* result = u64str(amount);
+  char tmp[MAX_DIGITS_IN_UINT64];
+  memcpy(tmp, result, sizeof(tmp));
+  explicit_bzero(result, sizeof(tmp));
+  adjustDecimals(tmp, strlen(tmp), result, MAX_DIGITS_IN_UINT64, decimals);
+  result[MAX_DIGITS_IN_UINT64-1] = '\0';
+  return result;
+}
