+++ src/ui_txn.c
@@ -8,10 +8,10 @@
 #include "algo_asa.h"
 #include "base64.h"
 #include "glyphs.h"
+#include "str.h"
 #include "ui_txn.h"
 
 #define CHECKEDSUM_BUFFER_SIZE 65
-#define MAX_DIGITS_IN_UINT64 27
 
 
 bool is_opt_in_tx(){
@@ -28,82 +28,6 @@ bool is_opt_in_tx(){
 }
 
 char caption[20];
-
-bool adjustDecimals(char *src, uint32_t srcLength, char *target,
-                    uint32_t targetLength, uint8_t decimals) {
-    uint32_t startOffset;
-    uint32_t lastZeroOffset = 0;
-    uint32_t offset = 0;
-    if ((srcLength == 1) && (*src == '0')) {
-        if (targetLength < 2) {
-                return false;
-        }
-        target[0] = '0';
-        target[1] = '\0';
-        return true;
-    }
-    if (srcLength <= decimals) {
-        uint32_t delta = decimals - srcLength;
-        if (targetLength < srcLength + 1 + 2 + delta) {
-            return false;
-        }
-        target[offset++] = '0';
-        target[offset++] = '.';
-        for (uint32_t i = 0; i < delta; i++) {
-            target[offset++] = '0';
-        }
-        startOffset = offset;
-        for (uint32_t i = 0; i < srcLength; i++) {
-            target[offset++] = src[i];
-        }
-        target[offset] = '\0';
-    } else {
-        uint32_t sourceOffset = 0;
-        uint32_t delta = srcLength - decimals;
-        if (targetLength < srcLength + 1 + 1) {
-            return false;
-        }
-        while (offset < delta) {
-            target[offset++] = src[sourceOffset++];
-        }
-        if (decimals != 0) {
-            target[offset++] = '.';
-        }
-        startOffset = offset;
-        while (sourceOffset < srcLength) {
-            target[offset++] = src[sourceOffset++];
-        }
-  target[offset] = '\0';
-    }
-    for (uint32_t i = startOffset; i < offset; i++) {
-        if (target[i] == '0') {
-            if (lastZeroOffset == 0) {
-                lastZeroOffset = i;
-            }
-        } else {
-            lastZeroOffset = 0;
-        }
-    }
-    if (lastZeroOffset != 0) {
-        target[lastZeroOffset] = '\0';
-        if (target[lastZeroOffset - 1] == '.') {
-                target[lastZeroOffset - 1] = '\0';
-        }
-    }
-    return true;
-}
-
-static char*
-amount_to_str(uint64_t amount, uint8_t decimals){
-  char* result = u64str(amount);
-  char tmp[MAX_DIGITS_IN_UINT64];
-  memcpy(tmp, result, sizeof(tmp));
-  explicit_bzero(result, sizeof(tmp));
-  adjustDecimals(tmp, strlen(tmp), result, MAX_DIGITS_IN_UINT64, decimals);
-  result[MAX_DIGITS_IN_UINT64-1] = '\0';
-  return result;
-}
-
 
 static void checksum_and_put_text(const uint8_t * buffer)
 {
